Update pog protocol to microsoft standard
