+++ src/ProofObligationPanel.ts
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 import { Uri, WebviewPanel, Disposable, window, ViewColumn, workspace, Webview } from 'vscode'
-import { ProofObligation } from "./protocol.slsp"
+import { ProofObligation } from "./protocol/proofObligationGeneration.slsp"
 import { Protocol2CodeConverter } from 'vscode-languageclient';
 import * as util from "./Util"
 import { createConverter } from 'vscode-languageclient/lib/common/protocolConverter';
